requirements.txt
streamlit
plotly
pillow
requests
google-generativeai
streamlit-chat
python-dotenv
PyMuPDF
